refactor: move refresh button
