+++ apps/mobile/src/components/atoms/Button/index.ts
@@ -1,2 +1,3 @@
 export * from './Button';
 export * from './IconButton';
+export * from './RefreshButton';
