+++ apps/mobile/src/screens/settings/ExternalServicesSettings.tsx
@@ -3,13 +3,7 @@ import React from 'react';
 import { ActionType, WatchListImportStatus } from '@naikamu/shared';
 import { formatDistanceToNow } from 'date-fns';
 import { pl } from 'date-fns/locale';
-import { Pressable, ScrollView, StyleSheet, Text, View } from 'react-native';
-import Animated, {
-  Easing,
-  useAnimatedStyle,
-  useSharedValue,
-  withTiming,
-} from 'react-native-reanimated';
+import { ScrollView, StyleSheet, Text, View } from 'react-native';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 import styled from 'styled-components/native';
 
@@ -24,6 +18,7 @@ import {
   Button,
   Link,
   PageLayout,
+  RefreshButton,
   SettingInputs,
   SettingsGroup,
   useLayout,
@@ -52,13 +47,9 @@ const getIconForStatus = (status: WatchListImportStatus) => {
   }
 };
 
-const duration = 1000;
-const easing = Easing.bezier(0.25, -0.5, 0.25, 1);
-
 export function ExternalServicesSettings({
   navigation,
 }: SettingsStackExternalServicesSettingsScreenProps) {
-  // const user = useUserStore(state => state.user);
   const { data: user } = useQueryUser();
   const { mutation } = useMutationSaveShindenUserId();
   const { data: watchListImportHistory, refetch: watchListImportRefetch } =
@@ -68,12 +59,6 @@ export function ExternalServicesSettings({
   const { translate } = useTranslate();
 
   const layout = useLayout();
-
-  const sv = useSharedValue<number>(0);
-
-  const animatedStyle = useAnimatedStyle(() => ({
-    transform: [{ rotate: `${sv.value * 360}deg` }],
-  }));
 
   return (
     <PageLayout.Default {...layout}>
@@ -142,19 +127,7 @@ export function ExternalServicesSettings({
             <Text style={[[colors.textLight, fontStyles.headerSmall]]}>
               {translate('settings.externalServices.lastImports')}
             </Text>
-            <Animated.View style={animatedStyle}>
-              <Pressable
-                onPress={() => watchListImportRefetch()}
-                onPressIn={() => {
-                  sv.value = withTiming(
-                    1,
-                    { duration, easing },
-                    () => (sv.value = 0),
-                  );
-                }}>
-                <Icon color={colors.textLight.color} name="refresh" size={30} />
-              </Pressable>
-            </Animated.View>
+            <RefreshButton refresh={watchListImportRefetch} />
           </Row>
           {user && user.shindenUserId && watchListImportHistory ? (
             watchListImportHistory.map(importHistory => (
